refinements to docs-gathering
