+++ filemagic.sh
@@ -31,4 +31,4 @@ apptainer exec \
     --bind "$PWD:$PWD" \
     --pwd "$PWD" \
     $SIF_NAME \
-    python3 scrape_carina.py
+    python3 scrape_static_docs.py
